Adapt imports to viewer subdir. Rename 'setup' classes to 'layout'.
